Create LS solver
A large commit for a large task: implementing the actual LS logic
taking it from nothing to a presentable state.

Everything is in solver.hpp
Prediction and Jacobian were tested in solver_stuff.cpp
The algorithm is tested in the aptly-named testone.cpp

Created a smaller dataset, mini*.g2o, by taking a small part of the
original dataset, to do some tests on a smaller problem.

Added support code in definitions and state.

Very very minor changes to some satellite code.